pyproject.toml
[project]
name = "data-analysis-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "load-dotenv>=0.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "tabulate>=0.9.0",
]
